+++ darts/models/forecasting/nlinear.py
@@ -110,11 +110,15 @@ class _NLinearModule(PLForecastingModule):
     def forward(self, x_in: PLModuleInput):
         """
         x_in
-            comes as tuple `(x_past, x_future, x_static)` where `x_past` is the input/past chunk and `x_future`
-            is the output/future chunk. Input dimensions are `(n_samples, n_time_steps, n_variables)`
+            comes as tuple `(x, x_future, x_static)` where `x` is the past target, past covariates and
+            historic future covariate chunk and `x_future` is the (non-historic) future chunk.
+            Input dimensions are `(n_samples, n_time_steps, n_variables)`
         """
         x, x_future, x_static = x_in  # x: (batch, in_len, in_dim)
+        # we clone `x`, to avoid value mutation from normalization when performing auto-regression
+        x = x.clone()
         batch, _, _ = x.shape
+        seq_last = None
 
         if self.shared_weights:
             # discard covariates, to ensure that in_dim == out_dim
@@ -122,7 +126,7 @@ class _NLinearModule(PLForecastingModule):
             x = x.permute(0, 2, 1)  # (batch, out_dim, in_len)
 
             if self.normalize:
-                seq_last = x[:, :, -1:].detach()  # (batch, out_dim, 1)
+                seq_last = x[:, :, -1:].detach().clone()  # (batch, out_dim, 1)
                 x = x - seq_last
 
             x = self.layer(x)  # (batch, out_dim, out_len * nr_params)
@@ -137,10 +141,11 @@ class _NLinearModule(PLForecastingModule):
             x = x.permute(0, 2, 1, 3)
         else:
             if self.normalize:
-                # get last values only for target features
-                seq_last = x[:, -1:, : self.output_dim].detach().clone()
-                # normalize the target features only (ignore the covariates)
-                x[:, :, : self.output_dim] = x[:, :, : self.output_dim] - seq_last
+                # get last values for all x but not future covariates
+                past_dim = self.input_dim - self.future_cov_dim
+                seq_last = x[:, -1:, :past_dim].detach().clone()
+                # normalize the input
+                x[:, :, :past_dim] -= seq_last
 
             x = self.layer(x.view(batch, -1))  # (batch, out_len * out_dim * nr_params)
             x = x.view(
@@ -168,10 +173,12 @@ class _NLinearModule(PLForecastingModule):
                     batch, self.output_chunk_length, self.output_dim * self.nr_params
                 )
 
-            x = x.view(batch, self.output_chunk_length, self.output_dim, self.nr_params)
             if self.normalize:
-                # model only forecasts target components, no need to slice
-                x = x + seq_last.view(seq_last.shape + (1,))
+                # Reverse the normalization for the target
+                x = x + seq_last[:, :, : self.output_dim]
+
+            x = x.view(batch, self.output_chunk_length, self.output_dim, self.nr_params)
+
         return x
 
 
@@ -183,7 +190,7 @@ class NLinearModel(MixedCovariatesTorchModel):
         output_chunk_shift: int = 0,
         shared_weights: bool = False,
         const_init: bool = True,
-        normalize: bool = False,
+        normalize: bool = True,
         use_static_covariates: bool = True,
         **kwargs,
     ):
@@ -227,7 +234,10 @@ class NLinearModel(MixedCovariatesTorchModel):
             initialization is used (default='True').
         normalize
             Whether to apply the simple "normalization" proposed in the paper, which consists
-            in subtracting the last value of the input sequence from the input sequence. Default: False.
+            in subtracting the last value of the input sequence from the input sequence. This is applied for the target
+            series and past covariates, but not future covariates because it would defeat the use of encoders (see
+            `add_encoders`). Without normalization, the models behaves like the simple Linear model proposed in the
+            paper. Default: True.
 
             .. note::
                 This cannot be applied to probabilistic models.