+++ darts/tests/models/forecasting/test_torch_forecasting_model.py
@@ -2100,6 +2100,10 @@ class TestTorchForecastingModel:
             length=model_kwargs["input_chunk_length"]
             + model_kwargs["output_chunk_length"]
             + 1
+        ) + tg.sine_timeseries(
+            length=model_kwargs["input_chunk_length"]
+            + model_kwargs["output_chunk_length"]
+            + 1
         )
         if not univ_series:
             ts = ts.stack(ts)